+++ src/loaders/OMETiffLoader.js
@@ -89,14 +89,18 @@ export default class OMETiffLoader {
    */
   _getIFDIndex({ z = 0, channel, time = 0 }) {
     let channelIndex;
+    // Without names, enforce a numeric channel indexing scheme
     if (this.channelNames.every(v => !v)) {
-      // Without names, enforce a numeric channel indexing scheme
       console.warn(
-        'There are no channel names in the OMEXML.  Please use numeric indexing'
+        'No channel names found in OMEXML.  Please be sure to use numeric indexing.'
       );
       channelIndex = channel;
-    } else {
+    } else if (typeof channel === 'string') {
       channelIndex = this.channelNames.indexOf(channel);
+    } else if (typeof channel === 'number') {
+      channelIndex = channel;
+    } else {
+      throw new Error('Channel selection must be numeric index or string');
     }
     const { SizeZ, SizeT, SizeC, DimensionOrder } = this.omexml;
     switch (DimensionOrder) {